# Conflicts:
#	client/src/components/Editor.js
#	client/src/hooks/hooks.js

+++ server/src/routes/api.js
@@ -1,22 +1,23 @@
 const fs = require("fs");
-const paths = require("../utils/utils");
-
+const path = require("path");
 const express = require("express");
 const router = express.Router();
+const paths = require("../utils/utils");
 
-router.delete("/deleteConverted", (req) => {
-  const filename = req.body.filename;
+router.delete("/deleteConverted", (req, res) => {
+  const filename = path.parse(req.body.filename).name;
   const format = req.body.format;
+  const videoPath = `${paths.basePath}/${paths.baseFolder}/${paths.video.folder}/${filename}.${format}`;
+  const thumbPath = `${paths.basePath}/${paths.baseFolder}/${paths.thumb.folder}/${filename}.jpg`;
+
   try {
-    fs.unlinkSync(
-      `${paths.basePath}/${paths.baseFolder}/${paths.video.folder}/${filename}.${format}`
-    );
-    fs.unlinkSync(
-      `${paths.basePath}/${paths.baseFolder}/${paths.thumb.folder}/${filename}.jpg`
-    );
-    //file removed
-  } catch (err) {
-    console.log(err);
+    if (fs.existsSync(videoPath)) fs.unlinkSync(videoPath);
+    if (fs.existsSync(thumbPath)) fs.unlinkSync(thumbPath);
+    res.json({ success: true });
+  } catch (e) {
+    console.error(e);
+    res.status(500);
+    res.json({ success: false, message: e.message });
   }
 });
